+++ src/meltano/api/controllers/m5o_file_parser.py
@@ -1,6 +1,9 @@
 from pyhocon import ConfigFactory
 from pathlib import Path
 from jinja2 import Template
+from typing import Dict, List
+
+import logging
 import json
 
 
@@ -112,6 +115,7 @@ class MeltanoAnalysisFileParser:
             conf = self.parse_m5o_file(model)
             parsed_model = self.model(conf, file_name)
             self.models.append(parsed_model)
+
         return self.models
 
     def model(self, ma_file_model_dict, file_name):
@@ -172,8 +176,9 @@ class MeltanoAnalysisFileParser:
         for join_name, join_def in ma_file_joins_dict.items():
             temp_join = {}
             temp_join["name"] = join_name
+            related_table_name = join_def.get("from", join_name)
             matching_table = self.table_conf_by_name(
-                temp_join["name"], "join", "name", file_name
+                related_table_name, "join", "name", file_name
             )
             temp_join["related_table"] = self.table(
                 self.parse_m5o_file(matching_table), matching_table.parts[-1]
@@ -212,28 +217,17 @@ class MeltanoAnalysisFileParser:
             raise MeltanoAnalysisFileParserMissingFieldsError(
                 missing_properties, "table", file_name
             )
+
         for prop_name, prop_def in table_file.items():
-            temp_table[prop_name] = prop_def
-            if prop_name == "columns":
-                temp_table[prop_name] = self.columns(prop_def)
-            elif prop_name == "aggregates":
-                temp_table[prop_name] = self.aggregates(prop_def)
+            subproperties = ("columns", "aggregates", "timeframes")
+
+            if prop_name in subproperties:
+                temp_table[prop_name] = self.name_flatten_dict(prop_def)
+            else:
+                temp_table[prop_name] = prop_def
+
         return temp_table
 
-    def columns(self, ma_file_columns_dict):
-        temp_columns = []
-        for column_name, column_def in ma_file_columns_dict.items():
-            temp_column = {"name": column_name}
-            for prop_name, prop_def in column_def.items():
-                temp_column[prop_name] = prop_def
-            temp_columns.append(temp_column)
-        return temp_columns
-
-    def aggregates(self, ma_file_aggregates_dict):
-        temp_aggregates = []
-        for aggregate_name, aggregate_def in ma_file_aggregates_dict.items():
-            temp_aggregate = {"name": aggregate_name}
-            for prop_name, prop_def in aggregate_def.items():
-                temp_aggregate[prop_name] = prop_def
-            temp_aggregates.append(temp_aggregate)
-        return temp_aggregates
+    @classmethod
+    def name_flatten_dict(cls, d: Dict) -> List:
+        return [{"name": k, **rest} for k, rest in d.items()]