+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 * Updated docs with `.m5o` authoring requirements and examples
+* add support for timeframes in tables
+
 ### Changes
 
 ### Fixes
@@ -55,6 +57,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 * update `meltano elt` to auto generate the job_id (job_id has been converted to an optional argument)
 
 ### Fixes
+* left joins now work correctly in analyze.
 * fixed broken sql toggles in analyze view
 * fixed sql output based on sql toggles in analyze view